chore(budgets/generators): move budgets workflow random to generators for dev pourposes
