+++ decidim-generators/lib/decidim/generators/app_templates/budgets_initializer.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require "budgets_workflow_random"
+Decidim::Budgets.workflows[:random] = "BudgetsWorkflowRandom"
